Don't let powershell treat bold as bright
Previously, ssh-ing from powershell results in some issues for base16
color themes. The issue was that powershell itself was converting bold
text to 'bright' colors before the result ever got to Alacritty. This is
a bit of a hack to undo that. I doubt this is the right long-term
change, but I've been using it for a few months and it works great.